requirements.txt
flask==3.0.3
google-cloud-bigquery==3.25.0
google-auth==2.30.0
beautifulsoup4==4.12.3
